On Master using locally hosted server I get > 3s response times:
```shell
> time curl -s 'localhost:8080/api/preprints-no-content?limit=100&page=1' > /dev/null

real	0m2.982s
user	0m0.020s
sys	0m0.054s

[07:27:59] 2s 0  ✔️
> time curl -s 'localhost:8080/api/preprints-no-content?limit=100&page=2' > /dev/null

real	0m3.103s
user	0m0.020s
sys	0m0.054s

[07:28:04] 3s 0  ✔️
> time curl -s 'localhost:8080/api/preprints-no-content?limit=100&page=3' > /dev/null

real	0m3.024s
user	0m0.019s
sys	0m0.048s

[07:28:11] 3s 0  ✔️
```

With this branch, we remove more superflous content, and apply pagination ealier. This results in significant speed increase:

```shell
> time curl -s 'localhost:8080/api/preprints-no-content?limit=100&page=1' > /dev/null

real	0m0.569s
user	0m0.006s
sys	0m0.012s

[07:47:10] 577ms 0  ✔️
> time curl -s 'localhost:8080/api/preprints-no-content?limit=100&page=2' > /dev/null

real	0m0.548s
user	0m0.011s
sys	0m0.020s

[07:47:13] 544ms 0  ✔️
> time curl -s 'localhost:8080/api/preprints-no-content?limit=100&page=3' > /dev/null

real	0m0.523s
user	0m0.011s
sys	0m0.019s

[07:47:16] 517ms 0  ✔️
```